backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi[standard]>=0.136.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
